optimized some replace utils
fix add rangeofmarker

fix function get hasecode

format code

+++ src/Nncase.Tests/Rules/Neutral/UnitTestFoldConv2DAddMul.cs
@@ -81,11 +81,19 @@ public class UnitTestFoldConv2DAddMul : TestClassBase
                           IR.F.Random.Normal(DataTypes.Float32, 0, 1, 1, new[] { 64, 3, 7, 7 }).Evaluate().AsTensor(),
                           IR.F.Random.Normal(DataTypes.Float32, 0, 1, 1, new[] { 64 }).Evaluate().AsTensor(),
                           new[] { 2, 2 },
-                          new[,] { { 3, 3 }, { 3, 3 }, },
+                          new[,]
+                            {
+                                { 3, 3 },
+                                { 3, 3 },
+                            },
                           new[] { 1, 1 }, PadMode.Constant, 1,
                           new[] { 0.0f, 6.0f }); // f32[1,64,112,112]
 
-            var v2 = IR.F.NN.ReduceWindow2D(ReduceOp.Mean, v1, (float)-3.4028235E+38, new[] { 3, 3 }, new[] { 2, 2 }, new[,] { { 1, 1 }, { 1, 1 }, }, new[] { 1, 1 }, false, false); // f32[1,64,56,56]
+            var v2 = IR.F.NN.ReduceWindow2D(ReduceOp.Mean, v1, -3.4028235E+38F, new[] { 3, 3 }, new[] { 2, 2 }, new[,]
+            {
+                { 1, 1 },
+                { 1, 1 },
+            }, new[] { 1, 1 }, false, false); // f32[1,64,56,56]
             var v3 = IR.F.Math.Mul(v2, IR.F.Random.Normal(DataTypes.Float32, 0, 1, 0, new[] { 64, 1, 1 })); // f32[1,64,56,56]
             var v4 = IR.F.Math.Add(v3, IR.F.Random.Normal(DataTypes.Float32, 0, 1, 0, new[] { 64, 1, 1 })); // f32[1,64,56,56]
             var v5 = IR.F.NN.Relu(v4); // f32[1,64,56,56]
@@ -94,7 +102,11 @@ public class UnitTestFoldConv2DAddMul : TestClassBase
                           IR.F.Random.Normal(DataTypes.Float32, 0, 1, 1, new[] { 256, 64, 1, 1 }).Evaluate().AsTensor(),
                           IR.F.Random.Normal(DataTypes.Float32, 0, 1, 1, new[] { 256 }).Evaluate().AsTensor(),
                           new[] { 1, 1 },
-                          new[,] { { 0, 0 }, { 0, 0 }, },
+                          new[,]
+                            {
+                                { 0, 0 },
+                                { 0, 0 },
+                            },
                           new[] { 1, 1 }, PadMode.Constant, 1,
                           new[] { 0.0f, 6.0f }); // f32[1,256,56,56]
             rootPre = v6;
